feat(Edit Mode) #30799 : Creating a REST Endpoint that abstract the Lucene Query generation and return results
