+++ dotCMS/src/main/java/com/dotcms/graphql/datafetcher/ContentMapDataFetcher.java
@@ -1,29 +1,30 @@
 package com.dotcms.graphql.datafetcher;
 
-import static com.dotmarketing.portlets.contentlet.transform.strategy.RenderFieldStrategy.isFieldRenderable;
-import static com.dotmarketing.portlets.contentlet.transform.strategy.RenderFieldStrategy.renderFieldValue;
-import static com.dotmarketing.portlets.contentlet.transform.strategy.TransformOptions.RENDER_FIELDS;
-
 import com.dotcms.contenttype.model.field.Field;
 import com.dotcms.graphql.DotGraphQLContext;
-import com.dotmarketing.util.json.JSONObject;
-import com.dotcms.rest.ContentResource;
+import com.dotcms.rest.ContentHelper;
 import com.dotmarketing.portlets.contentlet.model.Contentlet;
 import com.dotmarketing.portlets.contentlet.transform.DotContentletTransformer;
 import com.dotmarketing.portlets.contentlet.transform.DotTransformerBuilder;
 import com.dotmarketing.util.Logger;
 import com.dotmarketing.util.UtilMethods;
+import com.dotmarketing.util.json.JSONObject;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.liferay.portal.model.User;
 import graphql.Scalars;
 import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
 import io.vavr.control.Try;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Collectors;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+
+import static com.dotmarketing.portlets.contentlet.transform.strategy.RenderFieldStrategy.isFieldRenderable;
+import static com.dotmarketing.portlets.contentlet.transform.strategy.RenderFieldStrategy.renderFieldValue;
+import static com.dotmarketing.portlets.contentlet.transform.strategy.TransformOptions.RENDER_FIELDS;
 
 /**
  * {@link DataFetcher} that fetches the data when the special <code>_map<code/> GraphQL Field is requested.
@@ -82,7 +83,7 @@ public class ContentMapDataFetcher implements DataFetcher<Object> {
 
             // this only adds relationships to any json. We would need to return them with the transformations already
 
-            final JSONObject jsonWithRels = ContentResource.addRelationshipsToJSON(request, response,
+            final JSONObject jsonWithRels = ContentHelper.getInstance().addRelationshipsToJSON(request, response,
                     render.toString(), user, depth, false, contentlet,
                     contentMapInJSON, null, 1, true, false,
                     true);
